Refactor module access check to its own resolver.
